requirements.txt
absl-py==2.0.0
addict==2.4.0
aenum==3.1.15
affine==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
albumentations==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
antlr4-python3-runtime==4.9.3
anyio==4.1.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
biopython==1.81
bitsandbytes==0.41.3.post2
black==23.11.0
bleach==6.1.0
blessed==1.20.0
boto3==1.34.15
botocore==1.34.15
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
crc32c==2.4
crcmod==1.7
croniter==1.3.15
cryptography==41.0.7
cycler==0.12.1
dateutils==0.6.12
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.7.1
defusedxml==0.7.1
deprecation==2.1.0
docker-pycreds==0.4.0
docstring-parser==0.15
domdf-python-tools==3.7.0
editor==1.6.5
efficientnet-pytorch==0.7.1
einops==0.7.0
enum-tools==0.11.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.88.0
fastjsonschema==2.19.0
filelock==3.13.1
fiona==1.9.5
flake8==6.1.0
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
fvcore==0.1.5.post20221221
geopandas==0.14.3
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.25.1
google-auth-oauthlib==1.1.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.20.1
hydra-core==1.3.2
idna==3.6
imageio==2.33.1
importlib-metadata==6.9.0
importlib-resources==6.1.1
iniconfig==2.0.0
inquirer==3.2.1
iopath==0.1.10
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.14
jsonargparse==4.27.1
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.24.0
kiwisolver==1.4.5
kornia==0.7.0
laspy==2.5.1
lazy_loader==0.3
lightly==1.4.24
lightly-utils==0.0.2
lightning==2.1.3
lightning-cloud==0.5.57
lightning-utilities==0.10.0
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mmcv==2.1.0
mmengine==0.10.1
model-index==0.1.11
multidict==6.0.4
munch==4.0.0
mypy-extensions==1.0.0
natsort==8.4.0
natten==0.14.6+torch1130cu117
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
netCDF4==1.6.5
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.2
openxlab==0.0.29
ordered-set==4.1.0
oss2==2.17.0
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.4.0
pooch==1.8.0
portalocker==2.8.2
pretrainedmodels==0.7.4
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.13
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
PyMassSpec==2.3.0
pymzml==2.5.2
pyparsing==3.1.1
pyproj==3.6.1
pystac==1.9.0
pyteomics==4.6.3
pytest==8.0.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.1.3
pytz==2023.3.post1
pyvista==0.43.1
PyYAML==6.0.1
pyzmq==25.1.1
qudida==0.0.4
radiant-mlhub==0.4.1
rarfile==4.1
rasterio==1.3.9
readchar==4.0.5
referencing==0.31.1
regex==2023.10.3
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.13.2
rsa==4.9
Rtree==1.1.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
runs==1.2.0
s3transfer==0.10.0
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
scooby==0.9.2
sdjson==0.4.0
seaborn==0.13.0
segmentation-models-pytorch==0.3.3
Send2Trash==1.8.2
sentry-sdk==1.43.0
setproctitle==1.3.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.5
stack-data==0.6.3
starlette==0.22.0
starsessions==1.3.0
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminado==0.18.0
tfrecord==1.14.4
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.9.2
tinycss2==1.2.1
tokenize-rt==5.2.0
tomli==2.0.1
torch==1.13.1
torch-tb-profiler==0.4.3
torchmetrics==1.2.1
torchvision==0.14.1
tornado==6.4
tqdm==4.65.2
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typeshed-client==2.4.0
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.25.0
vtk==9.3.0
wandb==0.16.5
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.9
xmod==1.8.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipfile-deflate64==0.2.0
zipp==3.17.0
